Load app-conf by URL parameter
This introduces the ability to load a custom app-config JSON file, which
is defined by an URL parameter 'appCtx'.
The app-config has to be placed in the 'static' folder and the file has to be
named by the pattern app-conf-<appCtxParam>.json
If no param is given the default app-config is loaded (as before).