Add feature hover functionality to map
This adds a feature hover functionality on the map. By hovering a
feature of an activated vector layer a little tooltip-like element is shown on
the map as OL Overlay.
To activate a vector layer for hovering it has to be configured as 'hoverable'.
By setting a config option 'hoverAttribute' the shown attribute value
can be influenced.