Add image based rendering in InfoClickWin
This adds an (optional) additional image based rendering template in the
InfoClickWin component. This allows to show an image based and a
description based on configurable attributes of the clicked feature.
The template is (more or less) the one of the deprecated FeatureInfoWin,
which never made it as a module. Thus removes FeatureInfoWin component.